refactor: rename beta-flags as feature-flags in shared constants
