+++ shared/constants/index.ts
@@ -1,4 +1,4 @@
 export * from './file'
 export * from './form'
 export * from './links'
-export * from './beta-flags'
+export * from './feature-flags'
